Convert ThemedApp to React Testing Library
